Rename and move AsyncViaEventTest and *Messages
